# docs-site/openapi.yaml
openapi: 3.1.0
info:
  title: touchgrass API
  version: 1.0.0
  description: |
    Self-hosted blue-green deploys, metrics, and error tracking.
    Mirrors `internal/http/server.go` — update both together.

    All routes except `/api/health`, `/api/auth/login`,
    `/api/ingest`, and `/api/ingest/logs` require the admin session
    cookie from login.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api
    description: Same origin as the console
security:
  - cookieAuth: []
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Start an admin session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        '204': { description: Session cookie set }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /ingest:
    post:
      operationId: ingestReport
      tags: [ingest]
      summary: Submit one SDK report
      security: [{ keyAuth: [] }]
      description: |
        Public by design; the per-service key is the credential.
        Bodies cap at 1MB. Payloads are never logged.
      parameters:
        - $ref: '#/components/parameters/KeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestReport'
      responses:
        '202':
          description: Report evaluated
          content:
            application/json:
              schema:
                type: object
                required: [sampled]
                properties:
                  sampled:
                    type: boolean
                    description: False when the key's sampling rate dropped this report
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: Report exceeds 1MB }
  /ingest/logs:
    post:
      operationId: ingestLogBatch
      tags: [ingest]
      summary: Submit one batch of SDK log records
      security: [{ bearerAuth: [] }, { keyAuth: [] }]
      description: |
        Public by design; the per-service key is the credential, sent as
        `Authorization: Bearer <tg_key>` (or the `X-Touchgrass-Key` alias).
        Batches cap at 10MB and 1000 items. Validation is all-or-nothing:
        one invalid item rejects the batch with its index. Payloads are
        never logged.
      parameters:
        - name: Authorization
          in: header
          schema: { type: string, example: 'Bearer tg_0123456789abcdef0123456789abcdef' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SDKLogBatch'
      responses:
        '202':
          description: Batch stored
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted:
                    type: integer
                    description: Stored rows (the full batch; partial stores never happen)
        '400':
          description: Invalid batch or item
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/SDKLogItemError'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { description: Batch exceeds 10MB }
...
  /logs:
    get:
      operationId: searchLogs
      tags: [logs]
      summary: Search collected lines or SDK rows, newest first
      parameters:
        - $ref: '#/components/parameters/ServiceId'
        - name: source
          in: query
          description: Row source (default containers)
          schema: { type: string, enum: [containers, sdk], default: containers }
        - name: q
          in: query
          description: Full-text AND query (containers) or message substring (sdk)
          schema: { type: string }
        - name: trace_id
          in: query
          description: Full trace id filter (sdk only; 400 with source=containers)
          schema: { type: string, pattern: '^[0-9a-f]{32}$' }
        - name: after
          in: query
          schema: { type: string, format: date-time }
        - name: before
          in: query
          schema: { type: string, format: date-time }
        - name: stream
          in: query
          schema: { type: string, enum: [stdout, stderr] }
        - name: level
          in: query
          description: Comma-separated set (sdk adds trace and fatal)
          schema: { type: string, example: 'error,warn' }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lines, newest first
          content:
            application/json:
              schema:
                type: object
                required: [lines]
                properties:
                  lines:
                    type: array
                    description: LogLine rows for source=containers (levels parsed at read time), SDKLog rows for source=sdk
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/LogLine'
                        - $ref: '#/components/schemas/SDKLog'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: touchgrass_session
      description: Admin session from POST /api/auth/login
    keyAuth:
      type: apiKey
      in: header
      name: X-Touchgrass-Key
      description: Per-service tg_ ingestion key
    bearerAuth:
      type: http
      scheme: bearer
      description: Per-service tg_ ingestion key as a Bearer token
  parameters:
    ServiceId:
      name: service_id
      in: query
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 1000 }
    KeyHeader:
      name: X-Touchgrass-Key
      in: header
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or bad credential
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
...
    StackFrame:
      type: object
      required: [function, file, line, column]
      properties:
        function: { type: string }
        file: { type: string }
        line: { type: integer }
        column: { type: integer }
    Breadcrumb:
      type: object
      required: [at, category, message]
      properties:
        at: { type: string }
        category: { type: string }
        message: { type: string }
    IngestReport:
      type: object
      required: [type, message]
      properties:
        type: { type: string, enum: [exception, message] }
        message: { type: string }
        stack:
          type: array
          items: { $ref: '#/components/schemas/StackFrame' }
        breadcrumbs:
          type: array
          items: { $ref: '#/components/schemas/Breadcrumb' }
        release: { type: string }
...
    SDKLogAttribute:
      type: object
      required: [value, type]
      properties:
        value:
          type: [string, number, boolean]
          description: Must match the declared type (integer values are whole numbers)
        type: { type: string, enum: [string, integer, double, boolean] }
    SDKLogItem:
      type: object
      required: [timestamp, level, body]
      properties:
        timestamp:
          type: number
          description: Event time as unix seconds (float)
        level: { type: string, enum: [trace, debug, info, warn, error, fatal] }
        body: { type: string, minLength: 1, maxLength: 8192 }
        severity_number:
          type: integer
          minimum: 1
          maximum: 24
          description: OTel severity; inferred from level (1/5/9/13/17/21) when absent
        trace_id: { type: string, pattern: '^[0-9a-f]{32}$' }
        span_id: { type: string, pattern: '^[0-9a-f]{16}$' }
        attributes:
          type: object
          maxProperties: 64
          additionalProperties: { $ref: '#/components/schemas/SDKLogAttribute' }
    SDKLogBatch:
      type: object
      required: [items]
      properties:
        release: { type: string, maxLength: 128 }
        items:
          type: array
          minItems: 1
          maxItems: 1000
          items: { $ref: '#/components/schemas/SDKLogItem' }
    SDKLogItemError:
      type: object
      required: [error, index]
      properties:
        error: { type: string }
        index:
          type: integer
          description: Zero-based position of the invalid item in the batch
    SDKLog:
      type: object
      required: [id, service_id, ts, level, severity_number, message, attributes, trace_id, span_id, release, created_at]
      properties:
        id: { type: integer }
        service_id: { type: string }
        ts: { type: string, format: date-time }
        level: { type: string, enum: [trace, debug, info, warn, error, fatal] }
        severity_number: { type: integer, minimum: 1, maximum: 24 }
        message: { type: string }
        attributes:
          type: object
          additionalProperties: { $ref: '#/components/schemas/SDKLogAttribute' }
        trace_id: { type: string }
        span_id: { type: string }
        release: { type: string }
        created_at: { type: string, format: date-time }
    LogLine:
      type: object
      required: [id, service_id, container, stream, level, line, ts, created_at]
      properties:
        id: { type: integer }
        service_id: { type: string }
        container: { type: string }
        stream: { type: string, enum: [stdout, stderr] }
        level: { type: string, enum: [error, warn, info, debug] }
        line: { type: string }
        ts: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...